separated zenlink into module
